Updated Products

+++ app/controllers/users_controller.rb
@@ -1,11 +1,11 @@
 class UsersController < ApplicationController
   def index
-    redirect_to root_path, :alert => "Unauthorized"   unless current_user.admin?
+    redirect_to root_path, :alert => "Unauthorized" and return unless current_user.admin?
     @users = User.all.page(params[:page]).per(25).order('id ASC')
   end
 
   def show
-    redirect_to root_path, :alert => "Unauthorized" unless current_user.id == params[:id].to_i
+    redirect_to root_path, :alert => "Unauthorized" and return unless current_user.id == params[:id].to_i || current_user.admin?
     @user = User.find(params[:id])
   end
 
@@ -14,12 +14,12 @@ class UsersController < ApplicationController
   end
 
   def edit
-    redirect_to root_path, :alert => "Unauthorized" unless current_user.id == params[:id].to_i
+    redirect_to root_path, :alert => "Unauthorized" and return unless current_user.id == params[:id].to_i
     @user = User.find(params[:id])
   end
 
   def update
-    redirect_to root_path, :alert => "Unauthorized" unless current_user.id == params[:id].to_i
+    redirect_to root_path, :alert => "Unauthorized" and return unless current_user.id == params[:id].to_i || current_user.admin?
     @user = User.find(params[:id])
     if @user.update(user_params)
       redirect_to users_path
@@ -35,13 +35,14 @@ class UsersController < ApplicationController
     if @user.save
       log_in @user
       flash[:success] = "Welcome to the Webshop!"
-      redirect_to @user
+      redirect_to root_path
     else
       render 'new'
     end
   end
 
   def destroy
+    redirect_to root_path, :alert => "Unauthorized" and return unless current_user.admin?
     user = User.find(params[:id])
     user.destroy
     redirect_to users_path
